requirements.txt
django
pandas
matplotlib
scikit-learn
liac-arff
